pyproject.toml
[project]
name = "eia-data-backup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "pandas>=2.2.3",
    "path>=17.1.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "ruff>=0.9.5",
]
